pyproject.toml
[project]
name = "albert-tchap"
version = "0.2.1"
description = "Albert Tchap"
authors = [
    { name = "Pôle d'Expertise de la Régulation Numérique", email = "contact.peren@finances.gouv.fr" },
    { name = "Etalab", email = "etalab@modernisation.gouv.fr" },
]

requires-python = ">= 3.10"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "pillow==10.4.0",
    "markdown==3.6",
    "requests==2.32.3",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.4",
    "matrix-nio[e2e]==0.24.0",
    "cryptography==42.0.8",
    "structlog==24.2.0",
    "grist_api==0.1.0",
    "jinja2==3.1.4",
    "openai==1.51.0",
    "aiohttp==3.9.5",
]

# Packaging

[tool.setuptools]
py-modules = []

# Linter

[tool.ruff]
line-length = 100
src = ["app"]

[tool.ruff.isort]
known-third-party = ["nio", "matrix_bot"]

